Re-add report closer listener
Report display didn’t close: I renamed the function that adds its
listener, took it out of its IIFE and added a call to it from the
function that runs after reverseGeocode() - so now each time we update
the template, we get a new listener on the closer button. This should
probably eventually make its way into the pushHandlebars function
itself.